Added skhash: khash with a non-macro interface
skhash is a variant of khash with a non-macro programming interface.
The intention is to make khash more accessible as a library to use in
other programming languages, such as LuaJIT.

The hashtable parameterization is moved from C macros into hashtable
fields. The hope is that performance will be comparable. The hash and
equality functions are called indirectly (via function pointers) but
can still be specialized for specific key/value types and sizes.

skhash is written with relatively large keys and values in mind (up to
~64 bytes). I am not sure whether this is in line with the khash
design or whether this is totally the wrong idea and I should be
storing pointers in the table instead of whole key and values.

This implementation has only been trivially tested and has not been
benchmarked at all yet.